fsdp: one R3 test file per module under test
