+++ tests/fast/backends/test_fsdp_routing_replay.py
@@ -2,12 +2,14 @@ from tests.ci.ci_register import register_cpu_ci
 
 register_cpu_ci(est_time=60, suite="stage-a-cpu", labels=[])
 
+from argparse import Namespace
 from types import SimpleNamespace
 
 import pytest
 import torch.nn as nn
 
 from miles.backends.experimental.fsdp_utils.adaptations import routing_replay
+from miles.utils.arguments import resolve_fsdp_num_layers
 from miles.utils.replay_base import routing_replay_manager
 
 
@@ -48,6 +50,7 @@ def _reset_manager():
     routing_replay_manager.enabled = False
     routing_replay_manager.replays = []
     routing_replay_manager.current = None
+    routing_replay_manager.stage = routing_replay.FALLTHROUGH
 
 
 def test_discover_returns_global_layer_index_skipping_dense():
@@ -142,3 +145,126 @@ def test_install_raises_when_adapter_matches_but_finds_no_layers():
     )
     with pytest.raises(ValueError, match="found no MoE layers"):
         routing_replay.install(_model_with_layers(["moe"]), SimpleNamespace(model_type="empty"))
+
+
+def test_specs_register_adapters_for_every_supported_model_type():
+    import miles.backends.experimental.fsdp_utils.adaptations.specs  # noqa: F401
+
+    expected = {
+        "qwen3_moe": "Qwen3MoeTopKRouter",
+        "qwen3_5_moe_text": "Qwen3_5MoeTopKRouter",
+        "glm4_moe_lite": "Glm4MoeLiteMoE",
+    }
+    for model_type, module_cls_name in expected.items():
+        adapter = routing_replay.resolve_routing_replay_adapter(SimpleNamespace(model_type=model_type))
+        assert adapter is not None, model_type
+        assert adapter.module_cls_name == module_cls_name
+
+
+def test_dense_archs_do_not_resolve_to_a_moe_adapter():
+    import miles.backends.experimental.fsdp_utils.adaptations.specs  # noqa: F401
+
+    for model_type in ("qwen3_5_text", "qwen3"):
+        assert routing_replay.resolve_routing_replay_adapter(SimpleNamespace(model_type=model_type)) is None
+
+
+def test_enable_follows_use_routing_replay():
+    assert routing_replay.enable(Namespace(use_routing_replay=True, ci_test=False)) is True
+    assert routing_replay.enable(Namespace(use_routing_replay=False, ci_test=False)) is False
+
+
+def test_enable_defaults_false_when_arg_absent():
+    assert routing_replay.enable(Namespace(ci_test=False)) is False
+
+
+def test_enable_turns_on_the_replay_check_only_under_ci_test():
+    routing_replay.enable(Namespace(use_routing_replay=True, ci_test=True))
+    assert routing_replay_manager.enable_check_replay_result is True
+
+    routing_replay.enable(Namespace(use_routing_replay=True, ci_test=False))
+    assert routing_replay_manager.enable_check_replay_result is False
+
+
+def test_log_prob_stage_is_fallthrough_when_disabled():
+    routing_replay_manager.enabled = False
+    assert routing_replay.log_prob_stage(Namespace(use_rollout_routing_replay=True)) == routing_replay.FALLTHROUGH
+
+
+def test_log_prob_stage_replays_when_routing_came_from_the_rollout():
+    routing_replay_manager.enabled = True
+    assert routing_replay.log_prob_stage(Namespace(use_rollout_routing_replay=True)) == routing_replay.REPLAY_FORWARD
+
+
+def test_log_prob_stage_records_for_the_non_rollout_variant():
+    # --use-routing-replay alone has nothing to replay yet: the queues are only filled by fill(),
+    # which is gated on use_rollout_routing_replay. Replaying here would pop an empty queue.
+    routing_replay_manager.enabled = True
+    assert routing_replay.log_prob_stage(Namespace(use_rollout_routing_replay=False)) == routing_replay.RECORD
+
+
+def test_fill_is_skipped_for_the_non_rollout_variant():
+    # Passing None for every downstream argument: reaching fill_replay_data would raise.
+    routing_replay.fill(Namespace(use_rollout_routing_replay=False), None, None, None, None)
+
+
+def test_stage_sets_and_restores():
+    routing_replay_manager.stage = routing_replay.REPLAY_BACKWARD
+    with routing_replay.stage(routing_replay.REPLAY_FORWARD):
+        assert routing_replay_manager.stage == routing_replay.REPLAY_FORWARD
+    assert routing_replay_manager.stage == routing_replay.REPLAY_BACKWARD
+
+
+def test_stage_restores_on_exception():
+    routing_replay_manager.stage = routing_replay.REPLAY_BACKWARD
+    with pytest.raises(RuntimeError):
+        with routing_replay.stage(routing_replay.REPLAY_FORWARD):
+            raise RuntimeError("boom")
+    assert routing_replay_manager.stage == routing_replay.REPLAY_BACKWARD
+
+
+def test_stage_nests():
+    routing_replay_manager.stage = routing_replay.FALLTHROUGH
+    with routing_replay.stage(routing_replay.REPLAY_BACKWARD):
+        with routing_replay.stage(routing_replay.REPLAY_FORWARD):
+            assert routing_replay_manager.stage == routing_replay.REPLAY_FORWARD
+        assert routing_replay_manager.stage == routing_replay.REPLAY_BACKWARD
+    assert routing_replay_manager.stage == routing_replay.FALLTHROUGH
+
+
+def test_ref_model_creation_does_not_install_routing_replay():
+    import inspect
+
+    from miles.backends.experimental.fsdp_utils.actor import FSDPTrainRayActor
+
+    source = inspect.getsource(FSDPTrainRayActor._create_ref_model)
+    assert "routing_replay.install" not in source
+
+
+def test_resolve_num_layers_from_flat_config():
+    cfg = SimpleNamespace(num_hidden_layers=48)
+    assert resolve_fsdp_num_layers(cfg) == 48
+
+
+def test_resolve_num_layers_unwraps_text_config():
+    cfg = SimpleNamespace(text_config=SimpleNamespace(num_hidden_layers=40))
+    assert resolve_fsdp_num_layers(cfg) == 40
+
+
+def test_resolve_num_layers_prefers_text_config_when_both_present():
+    cfg = SimpleNamespace(num_hidden_layers=1, text_config=SimpleNamespace(num_hidden_layers=40))
+    assert resolve_fsdp_num_layers(cfg) == 40
+
+
+def test_resolve_num_layers_uses_get_text_config_when_available():
+    text = SimpleNamespace(num_hidden_layers=47)
+    cfg = SimpleNamespace(num_hidden_layers=1, get_text_config=lambda: text)
+    assert resolve_fsdp_num_layers(cfg) == 47
+
+
+def test_resolve_num_layers_falls_back_when_text_config_lacks_depth():
+    cfg = SimpleNamespace(num_hidden_layers=32, text_config=SimpleNamespace())
+    assert resolve_fsdp_num_layers(cfg) == 32
+
+
+def test_resolve_num_layers_returns_none_when_absent():
+    assert resolve_fsdp_num_layers(SimpleNamespace()) is None
